[REFACTOR] Create separate files for components in IssueList.js
[why]
Too many components were crammed into one source file.

[what]
IssueList.js
Components/IssueList/

[etc]
