+++ Frontend/Views/Components/IssueList.js
@@ -1,308 +1,9 @@
 import React, { useEffect, useState } from 'react';
 import { useHistory } from 'react-router-dom';
-import axios from 'axios';
-import useClickOutside from './Modal';
-
-const getUserId = (cookie) => {
-  const token = cookie.split('=')[1].split('.');
-  const { userId } = JSON.parse(window.atob(token[1]));
-  return userId;
-};
-
-const TopFilter = ({ reloadIssue, setResetQuery }) => {
-  const [isVisible, setIsVisible] = useState(false);
-  const history = useHistory();
-  const domNode = useClickOutside(() => {
-    setIsVisible(false);
-  });
-
-  const onToggleDropdown = () => {
-    setIsVisible(!isVisible);
-  };
-
-  const onClickFilter = (queryString) => {
-    history.push(`/issues?${queryString}`);
-    onToggleDropdown();
-    setResetQuery(true);
-    reloadIssue();
-  };
-
-  const TopFilterMenu = () => {
-    return (
-      <>
-        <div>Filter Issues</div>
-        <div onClick={() => onClickFilter('open=1')}>Open Issues</div>
-        <div onClick={() => onClickFilter(`author=${getUserId(document.cookie)}`)}>Your Issues</div>
-        <div onClick={() => onClickFilter(`assignee=${getUserId(document.cookie)}`)}>Everything assigned to you</div>
-        <div onClick={() => onClickFilter(`mentions=${getUserId(document.cookie)}`)}>Everything mentioning you</div>
-        <div onClick={() => onClickFilter('open=0')}>Closed Issues</div>
-      </>
-    );
-  };
-
-  return (
-    <>
-      <div ref={domNode}>
-        <button type="button" onClick={onToggleDropdown}>
-          Filters
-        </button>
-        <input type="text" />
-        {isVisible && <TopFilterMenu />}
-      </div>
-    </>
-  );
-};
-
-const MarkAs = ({ checkedIssues, reloadIssue, setIsMarkAs, setAllChecked }) => {
-  const [isVisible, setIsVisible] = useState(false);
-
-  const domNode = useClickOutside(() => {
-    setIsVisible(false);
-  });
-
-  const onToggleDropdown = () => {
-    setIsVisible(!isVisible);
-  };
-
-  const MarkAsList = () => {
-    const onChangeStatus = async (status) => {
-      try {
-        await axios.patch(
-          `${process.env.API_URL}/${process.env.API_VERSION}/issues/status`,
-          {
-            issues: checkedIssues,
-            isOpen: status,
-          },
-          { withCredentials: true },
-        );
-        onToggleDropdown();
-        setIsMarkAs(false);
-        setAllChecked(false);
-        reloadIssue();
-      } catch (err) {
-        console.log('error');
-      }
-    };
-
-    return (
-      <>
-        <div>Actions</div>
-        <div onClick={() => onChangeStatus(1)}>Open</div>
-        <div onClick={() => onChangeStatus(0)}>Closed</div>
-      </>
-    );
-  };
-
-  return (
-    <>
-      <div ref={domNode}>
-        <button type="button" onClick={onToggleDropdown}>
-          Mark as
-        </button>
-        {isVisible && <MarkAsList checkedIssues={checkedIssues} reloadIssue={reloadIssue} />}
-      </div>
-    </>
-  );
-};
-
-const UserItem = ({ value }) => {
-  return (
-    <>
-      <img src={value.imageLink} alt={value.username} />
-      <span>{value.username}</span>
-    </>
-  );
-};
-
-const LabelItem = ({ value }) => {
-  return (
-    <>
-      <span>{value.color}</span>
-      <span>{value.name}</span>
-      <span>{value.desc}</span>
-    </>
-  );
-};
-
-const MilestoneItem = ({ value }) => {
-  return (
-    <>
-      <span>{value.title}</span>
-    </>
-  );
-};
-
-const ChoiceList = ({ name, values, reloadIssue, onToggleDropdown, setResetQuery }) => {
-  let ItemComponent;
-  switch (name) {
-    case 'author':
-      ItemComponent = UserItem;
-      break;
-    case 'label':
-      ItemComponent = LabelItem;
-      break;
-    case 'milestone':
-      ItemComponent = MilestoneItem;
-      break;
-    case 'assignee':
-      ItemComponent = UserItem;
-      break;
-    default:
-      ItemComponent = null;
-  }
-
-  const history = useHistory();
-
-  const onSelectAlmaItem = (id) => {
-    const oldQueryString = window.location.search;
-
-    let oldQuery = {};
-    if (oldQueryString !== '') {
-      const querys = oldQueryString.replace('?', '').split('&');
-      oldQuery = querys.reduce((acc, query) => {
-        const [key, value] = query.split('=');
-        return { ...acc, [key]: value };
-      }, {});
-    }
-    const newQuery = {
-      ...oldQuery,
-      [name]: id,
-    };
-
-    const newQueryString = Object.keys(newQuery).reduce((acc, key) => {
-      return `${acc}${acc === '' ? '' : '&'}${key}=${newQuery[key]}`;
-    }, '');
-
-    history.push(`/issues?${newQueryString}`);
-    onToggleDropdown();
-    setResetQuery(true);
-    reloadIssue();
-  };
-
-  return (
-    <>
-      <div>{`Filter by ${name}`}</div>
-      {Object.keys(values).map((key) => {
-        return (
-          <div key={key} onClick={() => onSelectAlmaItem(key)}>
-            <span>@</span>
-            <ItemComponent value={values[key]} />
-          </div>
-        );
-      })}
-    </>
-  );
-};
-
-const Dropdown = ({ name, values, reloadIssue, setResetQuery }) => {
-  const [isVisible, setIsVisible] = useState(false);
-
-  const domNode = useClickOutside(() => {
-    setIsVisible(false);
-  });
-
-  const onToggleDropdown = () => {
-    setIsVisible(!isVisible);
-  };
-
-  return (
-    <>
-      <div ref={domNode}>
-        <button type="button" onClick={onToggleDropdown}>
-          {name}
-        </button>
-        {isVisible && (
-          <ChoiceList
-            name={name}
-            values={values}
-            reloadIssue={reloadIssue}
-            onToggleDropdown={onToggleDropdown}
-            setResetQuery={setResetQuery}
-          />
-        )}
-      </div>
-    </>
-  );
-};
-
-const Alma = ({ users, labels, milestones, reloadIssue, setResetQuery }) => {
-  return (
-    <>
-      <Dropdown name="author" values={users} reloadIssue={reloadIssue} setResetQuery={setResetQuery} />
-      <Dropdown name="label" values={labels} reloadIssue={reloadIssue} setResetQuery={setResetQuery} />
-      <Dropdown name="milestone" values={milestones} reloadIssue={reloadIssue} setResetQuery={setResetQuery} />
-      <Dropdown name="assignee" values={users} reloadIssue={reloadIssue} setResetQuery={setResetQuery} />
-    </>
-  );
-};
-
-const IssueListItem = ({
-  issue,
-  author,
-  labels,
-  assignees,
-  milestone,
-  setCheckedIssues,
-  checkedIssues,
-  isCheckAll,
-  setAllChecked,
-  allChecked,
-  isMarkAs,
-}) => {
-  const [isChecked, setIsChecked] = useState(false);
-
-  useEffect(() => {
-    setIsChecked(isCheckAll);
-  }, [isCheckAll]);
-
-  useEffect(() => {
-    if (allChecked) {
-      setIsChecked(allChecked);
-    }
-  }, [allChecked]);
-
-  useEffect(() => {
-    if (!isMarkAs) {
-      setIsChecked(false);
-    }
-  }, [isMarkAs]);
-
-  const onCheckIssue = () => {
-    if (isChecked === false) {
-      setIsChecked(true);
-      setCheckedIssues([...checkedIssues, issue.id]);
-    } else {
-      setIsChecked(false);
-      setCheckedIssues(checkedIssues.filter((elem) => elem !== issue.id));
-      setAllChecked(false);
-    }
-  };
-
-  return (
-    <div>
-      <br />
-      <div>
-        <input key={issue.id} checked={isChecked} type="checkbox" onChange={onCheckIssue} />
-        <span>{issue.title}</span>
-        <span>
-          {labels.map((label) => (
-            <span key={label.id}>{` ${label.name}`}</span>
-          ))}
-        </span>
-      </div>
-      <div>
-        <span>{`#${issue.id} `}</span>
-        <span>{issue.isOpen ? 'opened ' : 'closed '}</span>
-        <span>{`${issue.createdAt} by `}</span>
-        <span>{author.username}</span>
-      </div>
-      <div>{milestone && milestone.title}</div>
-      {assignees.map((assignee) => (
-        <span key={assignee.id}>{`@${assignee.username}`}</span>
-      ))}
-    </div>
-  );
-};
+import IssueListItem from './IssueList/IssueListItem';
+import TopFilter from './IssueList/TopFilter';
+import MarkAs from './IssueList/MarkAs';
+import Alma from './IssueList/Alma';
 
 const IssueList = ({ issues, users, labels, milestones, reloadIssue }) => {
   const [checkedIssues, setCheckedIssues] = useState([]);
